This enables us to run the execute() method in a JUnit test. I was finally able to reproduce the issue in testing.  That way I can run this in debug and figure out why the leading slash is not being put into referenced files after the fingerprint URL is substituted.
Added more configuration to the JUnit setup method to set private
fields in the fingerprintMojo itself.  We needed to set the sourceDirectory
and outputDirectory.

Added a sample URL to run through the fingerprinting process in  the
testing dummy file.  Also added the favicon.ico file to have something to
fingerprint.